--- frontend/.env_example
@@ -1,1 +1,0 @@
-DEPLOYED=true
