--- frontend/public/index.html
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Chess</title>
-    <link rel="stylesheet" href="style.css">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.5/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-SgOJa3DmI69IUzQ2PVdRZhwQ+dy64/BUtbMJw1MZ8t5HZApcHrRKUc4W0kG879m7" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.5/dist/js/bootstrap.bundle.min.js"></script>
-    <script defer type="text/javascript" src="client.js"></script>
-</head>
-
-<body>
-    <div class="container">
-        <main class="d-flex flex-column flex-md-row align-items-center align-items-md-start justify-content-center gap-3 w-100">
-            
-            <div id="board" class="grid-container"></div>
-            
-            <div class="d-flex flex-column gap-1 p-3 bg-light rounded shadow-sm w-100" style="max-width: 300px;">
-                
-                <div class="border-bottom pb-2">
-                    <strong class="mb-1 d-block">Material Count:</strong>
-                    <div class="d-flex justify-content-between">
-                        <div class="d-flex align-items-center">
-                            <span class="me-1">♔</span>
-                            <span>White:</span>
-                            <span id="whiteMaterial" class="ms-1">0</span>
-                        </div>
-                        <div class="d-flex align-items-center">
-                            <span class="me-1">♚</span>
-                            <span>Black:</span>
-                            <span id="blackMaterial" class="ms-1">0</span>
-                        </div>
-                    </div>
-                </div>
-
-                <div class="d-flex justify-content-between align-items-center border-bottom pb-2">
-                    <strong>Status:</strong>
-                    <span id="gameStatus">Normal</span>
-                </div>
-
-                 <div class="d-flex justify-content-between align-items-center border-bottom pb-2">
-                    <strong>Legal Moves:</strong>
-                    <span id="legalMoveCount">0</span>
-                </div>
-
-                <div class="d-flex justify-content-between align-items-center">
-                    <label class="form-check-label" for="autoqueenSwitch">
-                        <strong>Auto Queen</strong>
-                    </label>
-                    <div class="form-check form-switch">
-                        <input class="form-check-input" type="checkbox" role="switch" id="autoQueenSwitch" checked>
-                    </div>
-                </div>
-
-                <div class="d-flex flex-row gap-3 mt-2 justify-content-end">
-                    <div style="
-                    display: inline-flex;
-                    align-items: center;
-                    justify-content: center;
-                    border: 2px solid #333; 
-                    border-radius: 100%; 
-                    padding: 2px; ">   
-                    <strong>C_v2.4</strong>
-                </div>
-
-                    <button onclick="reset()" class="btn btn-danger">🔁 Reset</button>
-                    <button onclick="undo()" class="btn btn-outline-primary">↩️ Undo</button>
-                </div>
-            </div>
-        </main>
-    </div>
-
-    <div class="modal fade" id="promotionModal" tabindex="-1" aria-labelledby="promotionModalLabel" aria-hidden="true">
-        <div class="modal-dialog modal-dialog-centered">
-            <div class="modal-content text-center">
-            <div class="modal-header">
-                <h5 class="modal-title" id="promotionModalLabel">Choose Promotion Piece</h5>
-            </div>
-            <div class="modal-body d-flex justify-content-around">
-                <button class="btn btn-outline-primary promo-btn" data-piece="q">♕ Queen</button>
-                <button class="btn btn-outline-primary promo-btn" data-piece="r">♖ Rook</button>
-                <button class="btn btn-outline-primary promo-btn" data-piece="b">♗ Bishop</button>
-                <button class="btn btn-outline-primary promo-btn" data-piece="n">♘ Knight</button>
-            </div>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
