--- frontend/public/style.css
@@ -1,64 +1,0 @@
-.grid-container {
-    display: grid;
-    grid-template-columns: repeat(8, 1fr);
-    grid-template-rows: repeat(8, 1fr);
-    width: min(90vmin, 90vw, 600px);
-    aspect-ratio: 1 / 1;
-    background-image: url("/images/board.png");
-    background-size: cover;
-    background-repeat: no-repeat;
-}
-
-.grid-item {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 100%;
-    position: relative;
-}
-
-.grid-item img {
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-    display: block;
-}
-
-.red-outline {
-    outline: 2px solid red;
-    outline-offset: -4px;
-}
-
-.green-outline {
-    outline: 2px solid green;
-    outline-offset: -4px;
-}
-
-body {
-    margin: 0;
-    background-color: #f0f0f0;
-    min-height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 1rem;
-}
-
-.container {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-/* Additional responsive styles */
-@media (max-width: 768px) {
-    body {
-        padding: 0.5rem;
-    }
-    
-    .container {
-        padding: 0;
-    }
-}
