+++ frontend/src/components/Board.vue
