+++ frontend/src/components/Game.vue
