+++ frontend/src/components/Panel.vue
