--- vue/.gitignore
@@ -1,3 +1,0 @@
-.vscode
-.env
-/node_modules
