--- vue/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "frontend",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "bootstrap": "^5.3.8",
-    "dotenv": "^17.3.1",
-    "express": "^5.2.1",
-    "vue": "^3.5.28",
-    "vue-router": "^4.6.4",
-    "ws": "^8.19.0"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-vue": "^6.0.4",
-    "vite": "^7.3.1",
-    "vite-plugin-vue-devtools": "^8.0.6"
-  },
-  "engines": {
-    "node": "^20.19.0 || >=22.12.0"
-  }
-}
